Review approve part is added

+++ controllers/reviewController.js
@@ -45,13 +45,57 @@ export function getReview(req,res){
 export function deleteReview (req,res){
     const email = req.params.email;
 
-    Review.deleteOne({email : email}).then(()=>{
-        res.json({message : "Review deleted"})
-    }).catch(()=>{
-        res.status(500).json({
-            error : "Review deletion dailed"
+    if (req.user == null) {
+        res.status(401).json({message : "plaese loging and try again"});
+        return;
+    }
+
+    if (req.user.role == "admin"){
+        Review.deleteOne({email : email}).then(()=>{
+            res.json({message : "Review deleted"})
+        }).catch(()=>{
+            res.status(500).json({
+                error : "Review deletion dailed"
+            });
+        })
+        return;
+    }
+
+    if (req.user.role == "customer"){
+        if (req.user.email == email){
+            Review.deleteOne({email : email}).then(()=>{
+                res.json({message : "Review deleted"})
+            }).catch(()=>{
+                res.status(500).json({
+                    error : "Review deletion failed"
+                });
+            })
+        }else{
+            res.status(403).json({message : "you are not authorized to perform this action"});
+        }
+    }
+
+}
+
+export function approveReview(req,res){
+    const email = req.params.email;
+
+    if (req.user == null){
+        res.status(401).json({
+            message : "plaese login and try again"
         });
-    })
+        return;
+    }
+
+    if (req.user.role == "admin"){
+        Review.updateOne({email : email,},{isApproved : true,}).then(()=>{
+            res.json({message : "Review approved successfully"})
+        }).catch(()=>{
+            res.status(500).json({error : "Review approval faild"})
+        })
+    }else{
+        res.status(403).json({ message : "You are not an admin. only admin can approve the reviews "});
+    }
 }
